changed Seedings in GetTourney to use the regualr tourney seedings and do string splitting to get the seed. Changed the docstring to reflect that.
Added docstring to bracket_visualizer. made the bracket image file a parameter with a default value. Added a check to make sure the bracket_dict input contains all the necessary bracket slots, raises error if not and lists missing slots. Maybe I should move the whole function into GetTourney?

added bracket visualization to the main file in Implement Quotient to help with testing. also moved the location of the reg season detailed results to a parameter for init in the class with the old location as the default value